pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "psycopg[binary] (>=3.2.7,<4.0.0)",
    "whitenoise (>=6.9.0,<7.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
